requirements.txt
appdirs==1.4.2
<<<<<<< HEAD
argcomplete==1.8.2
asn1crypto==0.22.0
awsebcli==3.10.5
beautifulsoup4==4.5.3
blessed==1.14.2
boto==2.48.0
botocore==1.5.90
bs4==0.0.1
bz2file==0.98
cement==2.8.2
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
colorama==0.3.7
cryptography==2.0.3
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.1.2
dill==0.2.7.1
Django==1.11.3
django-appconf==1.0.2
django-autocomplete-light==3.2.9
django-bootstrap3==8.2.1
django-background-tasks==1.1.11
django-compressor==2.1.1
=======
awsebcli==3.10.5
beautifulsoup4==4.6.0
blessed==1.14.2
botocore==1.5.90
bs4==0.0.1
cement==2.8.2
colorama==0.3.7
Django==1.11.3
django-bootstrap3==8.2.1
>>>>>>> ded216f852c651889e7872ae31f367c57d02966f
django-crispy-forms==1.6.1
django-filter==1.0.4
django-filters==0.2.1
django-tables2==1.4.1
django-watson==1.3.1
django-widget-tweaks==1.4.1
doc2text==0.2.4
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
docx2txt==0.6
<<<<<<< HEAD
EbookLib==0.15
ftfy==4.4.3
future==0.16.0
fuzzywuzzy==0.15.1
gensim==2.3.0
html5lib==0.999999999
hypothesis==3.16.0
idna==2.6
image==1.5.11
jmespath==0.9.3
lxml==3.8.0
matplotlib==2.0.2
mime==0.1.0
murmurhash==0.26.4
mysqlclient==1.3.10
networkx==1.11
nltk==3.2.4
numpy==1.13.1
nxviz==0.3.1
olefile==0.44
packaging==16.8
palettable==3.0.0
pandas==0.20.3
pathlib==1.0.1
pathspec==0.5.0
Pillow==4.2.1
plac==0.9.6
polcart==2017.1.16.14.45
preshed==1.0.0
probableparsing==0.0.1
py==1.4.34
pycparser==2.18
pyparsing==2.2.0
PyPDF2==1.26.0
pytesseract==0.1.7
pytest==3.2.0
python-crfsuite==0.9.2
python-dateutil==2.6.1
python-Levenshtein==0.12.0
python-pptx==0.6.5
pytz==2017.2
PyYAML==3.12
rcssmin==1.0.6
regex==2017.7.28
requests==2.18.4
rjsmin==1.0.12
scipy==0.19.1
semantic-version==2.5.0
six==1.10.0
smart-open==1.5.3
spacy==1.9.0
tabulate==0.7.5
termcolor==1.1.0
thinc==6.5.2
toolz==0.8.2
tqdm==4.15.0
ujson==1.35
urllib3==1.22
usaddress==0.5.10
Wand==0.4.4
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.44.0
whitenoise==3.3.0
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==0.9.8
=======
future==0.16.0
image==1.5.11
jmespath==0.9.3
mime==0.1.0
mysqlclient==1.3.10
numpy==1.13.1
olefile==0.44
packaging==16.8
pathspec==0.5.0
Pillow==4.2.1
pyparsing==2.1.10
PyPDF2==1.26.0
pytesseract==0.1.7
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.9.1
scipy==0.19.1
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
wcwidth==0.1.7
websocket-client==0.44.0
whitenoise==3.3.0
>>>>>>> ded216f852c651889e7872ae31f367c57d02966f
